refactor: ball 클래스 재사용
